Backend pull request fixes (#13)
* Added Test for user fetch from principal

* Moved user fetch to own function

* Added "throws Exception"

* Re-added NoArgsConstructor

* Modified Payment-Get-Test

* Added Exception and Logger

* Replaced double with BigDecimal

* Replaced double with BigDecimal

* Replaced 0.0 with 0

* Fixed issue when changing payment

* Added initial spring test

+++ backend/src/main/java/com/github/redstylzz/backend/model/Category.java
@@ -7,6 +7,8 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -19,6 +21,6 @@ public class Category {
 
     String userID;
     String categoryName;
-    double paymentSum;
+    BigDecimal paymentSum;
 
 }
